enhancement: extract mac address from nic
When rewriting a ifcfg file, the value for HWADDR in the orignal ifcfg
file is reused.

However, it seems that there is a case that Anaconda doesn't recorded
HWADDR for nic other than eth0. In that case, newly generated ifcfg
file doesn't have HWADDR, too.

If a nic is not renamed, this does not a trouble. However, if a user
use `altname' that I introduced in the last commit, HWADDR is important
information for initscript finding association between a ifcfg file
and a nic.

This commit adds code to %post that extracts mac address from a nic
and records it to the generated ifcfg file for the nic as HWADDR only
if HWADDR is not given in the original ifcfg file.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>